Work in progress on a test setup for Byron;Shelley node testing
Modeled on the shelley-from-genesis setup script, but now for a
Byron;Shelley setup.

This is more complicated as we need a full Byron setup and most of a
Shelley setup too, and appropriately aligned between the two so we can
start block producing nodes in Byron mode and switch over to Shelley.